# Refactor `ULID.parse` with optimized `ULID.from_integer`