```
Warming up --------------------------------------
ULID.parse / After #7
                        12.036k i/100ms
ULID.parse2 / After #114
                        10.359k i/100ms
Calculating -------------------------------------
ULID.parse / After #7
                        112.137k (± 9.4%) i/s -    565.692k in   5.094086s
ULID.parse2 / After #114
                        104.315k (± 6.5%) i/s -    528.309k in   5.085027s

Comparison:
ULID.parse / After #7:   112136.8 i/s
ULID.parse2 / After #114:   104315.1 i/s - same-ish: difference falls within error
```

The result looks same, but it makes actually faster for use-cases, because `from_integer` cached the integer. ref: #112